+++ lib/compendium-api.rb
@@ -3,3 +3,5 @@ require 'nretnil/compendium-api/authentication'
 require 'nretnil/compendium-api/post'
 require 'nretnil/compendium-api/user'
 require 'nretnil/compendium-api/comment'
+
+require 'nretnil/compendium-helpers/helpers.rb'